+++ Optimize Info on Linux.md
@@ -227,3 +227,15 @@ nsys数据说明，最大的性能瓶颈在数据传输上
     施加多个同步点
     
     记录一下当前所需最小的同步点：在完成C的传输后一定要分流同步
+
+
+## v0.8
+现在最大的困难是传输困难，在这种情况下，只要传输的数据量不变，传输就难以优化
+所以，最直接的解决办法就是不要进行大量传输，而是在一开始的时候就进入gpu计算，从源头上解决问题
+开始的数据量是变换后数据量的1%，只要gpu中可以保持相同的性能，那么问题可以得到很好的解决
+
+尝试1：
+    transform分块处理：效率改变不大，核心原因是我没有规划cpu寄存器分配
+
+尝试2：
+    transform与packing合并处理
